converted datatype to float to accept decimal values as well

+++ basics/area of square.py
@@ -1,3 +1,3 @@
-a=int(input("Enter the value of side"))
+a=float(input("Enter the value of side"))
 b=a**2
 print("The area of square is",b)
